DataCiteConnector: don't rely on handles.
While the DOIIdentifierProvider checks our database to ensure that a DOI
is not reserverd/registered for another DSpaceObject.
The DataCiteConnector was checking this again agains  the meta data we
send to DataCite. Within these meta data we are submitting the handle as
alternative identifier and used it to check whether a DOI is
reserved/registered for a specific DSpaceObject. This commit removes
the checks from the DataCiteConnector for the following reasons:
 * In case we want to completly replace handles by DOIs one day, we
   shouldn't rely on handles. Further ongoing: an IdentifierProvider
   shouldn't rely on other identifiers than the one he provides.
 * If Item Level Versioning is enabled the handles may change depending
   on the VersionedHandleIdentifier that is used. This makes handles
   unreliable for the checks we were using them before.
   handles.